requirements.txt
fastapi
uvicorn
pydantic[email]
python-jose[cryptography]
bcrypt
pymongo
email-validator
